+++ src/main/java/org/folio/sidecar/model/error/ErrorCode.java
@@ -19,7 +19,8 @@ public enum ErrorCode {
   AUTHORIZATION_ERROR("authorization_error"),
   READ_TIMEOUT_ERROR("read_timeout_error"),
   ENTITLEMENTS_NOT_LOADED_ERROR("entitlements_not_loaded_error"),
-  MOD_USERS_KEYCLOAK_TARGET_NOT_RESOLVED_ERROR("mod_users_keycloak_target_not_resolved_error");
+  MOD_USERS_KEYCLOAK_TARGET_NOT_RESOLVED_ERROR("mod_users_keycloak_target_not_resolved_error"),
+  ROUTES_NOT_INITIALIZED_ERROR("routes_not_initialized_error");
 
   @JsonValue
   private final String value;